+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # CHANGELOG
 
+* `2.0.90` Release - [2.0.90](#2090)
 * `2.0.89` Release - [2.0.89](#2089)
 * `2.0.86` Release - [2.0.86](#2086)
 * `2.0.85` Release - [2.0.85](#2085)
@@ -13,10 +14,15 @@
 * `0.2.x` Releases - [0.2.0](#020) | [0.2.1](#021) | [0.2.2](#022)
 * `0.1.x` Releases - [0.1.0](#010) | [0.1.1](#011) | [0.1.2](#012) | [0.1.3](#013) | [0.1.4](#014) | [0.1.5](#015) | [0.1.6](#016) | [0.1.7](#017) | [0.1.8](#018) | [0.1.9](#019)
 
+## 2.0.90
+
+#### Features
+* Custom `RequestId`  on  `Dispatcher` used when making network requests.
+
 ## 2.0.89
 
 #### Changes
-`Dispatcher` resilience improvements.
+* `Dispatcher` resilience improvements.
 
 #### Bug Fixes
 * `EMP-11523` Fixed issue where `Device.type` would be incorrectly set while logging in.